requirements.txt
instaloader==4.9.6
PyYAML==6.0.1